requirements.txt
google-generativeai
pandas
Pillow
pytesseract
PyPDF2
python-docx
streamlit
reportlab